refactor: Change DefaultFactory inner workings to use same logic as PropertyMapper
