+++ tests/Implementation/Php81/Foo/BarItem.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace JsonMapper\Tests\Implementation\Php81\Foo;
 
-
 class BarItem
 {
     public function __construct(
